backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pymysql==1.1.0
cryptography==42.0.2
paho-mqtt==1.6.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
